fix observer attachment to gui buttons
These changes have removed redundent files for the gui that are no longer used or referenced, as well as allowing the new gui class system to work with observers. This has been achieved by using the built in method calls when subclassing the GTK Application class.
