--- src/MainApplication.h
@@ -1,31 +1,0 @@
-#ifndef _APPLICATION_H
-#define _APPLICATION_H
-
-#include <gtkmm-4.0/gtkmm.h>
-
-/**
- * Author: @michael-bailey
- * # MainApplication
- * This class defines the main application of the client
- * It handles all application level events like activation and opening files
- */
-class MainApplication : public Gtk::Application {
- private:
-  Glib::RefPtr<Gtk::Builder> builder;
-  Gtk::Window* main_window;
-
- protected:
-  MainApplication();
-  ~MainApplication();
-
-	// application lifecycle signals
-	void on_startup() override;
-  void on_activate() override;
-	void on_shutdown() override;
-	
-
- public:
-  static Glib::RefPtr<MainApplication> create();
-};
-
-#endif
