+++ src/controller/addfriend-observer.cc
@@ -3,15 +3,12 @@
 #include <iostream>
 #include <string>
 
-#include "../view/view.h"
-
-using namespace chat_client_view;
 using namespace chat_client_controller;
 using namespace chat_client_model;
 using namespace std;
 
 AddFriendObserver::AddFriendObserver(shared_ptr<ClientModel> model,
-                                     shared_ptr<View> view) {
+                                     Glib::RefPtr<MainApplication> view) {
   this->model = model;
   this->view = view;
