+++ src/controller/client-controller.h
@@ -2,13 +2,13 @@
 #define CPPCHATCLIENT_CONTROLLER_CLIENTCONTROLLER_H_
 
 #include "../model/client-model.h"
-#include "../view/view.h"
+#include "../view/MainApplication.h"
 
 namespace chat_client_controller {
 class ClientController {
  private:
   std::shared_ptr<chat_client_model::ClientModel> model;
-  std::shared_ptr<chat_client_view::View> view;
+  Glib::RefPtr<MainApplication> view;
 
  public:
   ClientController(int, char **);
