+++ src/controller/observer.h
@@ -1,7 +1,9 @@
 #ifndef CPPCHATCLIENT_CONTROLLER_OBSERVER_H_
 #define CPPCHATCLIENT_CONTROLLER_OBSERVER_H_
 
-#include <memory>
+#include <gtkmm-4.0/gtkmm.h>
+
+class MainApplication;
 
 namespace chat_client_view {
 class View;
@@ -10,7 +12,7 @@ class View;
 namespace chat_client_controller {
 class Observer {
  protected:
-  std::shared_ptr<chat_client_view::View> view;
+  Glib::RefPtr<MainApplication> view;
 
  public:
   virtual void Execute() = 0;
