+++ src/controller/view-factory.cc
@@ -1,16 +1,10 @@
 #include "view-factory.h"
 
-#include "../view/gui-view.h"
-#include "../view/view.h"
-
 using namespace chat_client_controller;
-using namespace chat_client_view;
 using namespace std;
 
-shared_ptr<View> ViewFactory::CreateView(int argc, char **argv) {
-  GuiView gui_view(argc, argv);
-  gui_view.Setup(argc, argv);
+Glib::RefPtr<MainApplication> ViewFactory::CreateView(int argc, char **argv) {
+  auto app = MainApplication::create();
 
-  shared_ptr<View> gui_ptr(new GuiView(move(gui_view)));
-  return gui_ptr;
+  return app;
 }
